refactor(runtime): extract direct policy contracts
